refactor(WebAPI): Moved GetAccessiblePlexServersCommand to vertical slice
