+++ src/PlexApi/PlexApiService.cs
@@ -163,115 +163,4 @@ public class PlexApiService : IPlexApiService
             plexServerConnection.Value
         );
     }
-
-    /// <inheritdoc />
-    public async Task<Result<List<PlexServerAccessDTO>>> GetAccessiblePlexServersAsync(int plexAccountId)
-    {
-        var plexAccount = await _dbContext.PlexAccounts.GetAsync(plexAccountId);
-        if (plexAccount is null)
-        {
-            return ResultExtensions.EntityNotFound(nameof(PlexAccount), plexAccountId);
-        }
-
-        var plexAccountToken = await GetPlexApiTokenAsync(plexAccount);
-        if (plexAccountToken.IsFailed)
-            return plexAccountToken.ToResult();
-
-        var result = await _plexApiWrapper.GetAccessibleServers(plexAccountToken.Value);
-        if (result.IsFailed)
-        {
-            return result.ToResult().LogError();
-        }
-
-        var plexServers = result
-            .Value.FindAll(x => x.Provides.Contains("server"))
-            .Select(x => new PlexServerAccessDTO
-            {
-                AccessToken = new ServerAccessTokenDTO
-                {
-                    PlexAccountId = plexAccountId,
-                    MachineIdentifier = x.ClientIdentifier,
-                    AccessToken = x.AccessToken,
-                    IsServerOwned = x.Owned,
-                },
-                PlexServer = new PlexServer
-                {
-                    Id = 0,
-                    Name = x.Name,
-
-                    // The servers have an OwnerId of 0 when it belongs to the PlexAccount that was used to request it.
-                    OwnerId = x.OwnerId ?? plexAccount.PlexId,
-                    PlexServerOwnerUsername = x.SourceTitle ?? plexAccount.Username,
-                    Device = x.Device ?? string.Empty,
-                    Platform = x.Platform ?? string.Empty,
-                    PlatformVersion = x.PlatformVersion ?? string.Empty,
-                    Product = x.Product,
-                    ProductVersion = x.ProductVersion,
-                    Provides = x.Provides,
-                    CreatedAt = x.CreatedAt,
-                    LastSeenAt = x.LastSeenAt,
-                    MachineIdentifier = x.ClientIdentifier,
-                    PublicAddress = x.PublicAddress,
-                    PreferredConnectionId = 0,
-                    IsEnabled = !_serverSettingsModule.GetIsHidden(x.ClientIdentifier),
-                    Home = x.Home,
-                    Synced = x.Synced,
-                    Relay = x.Relay,
-                    Presence = x.Presence,
-                    HttpsRequired = x.HttpsRequired,
-                    PublicAddressMatches = x.PublicAddressMatches,
-                    DnsRebindingProtection = x.DnsRebindingProtection,
-                    NatLoopbackSupported = x.NatLoopbackSupported,
-                    PlexAccountServers = [],
-                    PlexLibraries = [],
-                    ServerStatus = [],
-                    PlexServerConnections = x
-                        .Connections.Select(y => new PlexServerConnection
-                        {
-                            Id = 0,
-                            Protocol = y.Protocol.ToString().ToLower(),
-                            Address = y.Address,
-                            Port = y.Port,
-                            Local = y.Local,
-                            Relay = y.Relay,
-                            IPv4 = y.Address.IsIpAddress() && !y.IPv6,
-                            IPv6 = y.IPv6,
-                            Url = y.Uri,
-                            PlexServer = null,
-                            PlexServerId = 0,
-                            LatestConnectionStatus = null,
-                            IsCustom = false,
-                        })
-                        .ToList(),
-                },
-            })
-            .ToList();
-
-        return Result.Ok(plexServers);
-    }
-
-    private async Task<Result<string>> GetPlexApiTokenAsync(PlexAccount? plexAccount)
-    {
-        if (plexAccount == null)
-            return ResultExtensions.IsNull(nameof(plexAccount));
-
-        if (plexAccount.GetAuthToken != string.Empty)
-        {
-            // TODO:Make the token refresh limit configurable
-            if ((plexAccount.ValidatedAt - DateTime.UtcNow)?.TotalDays < 30)
-            {
-                _log.InformationLine("Plex AuthToken was still valid, using from local DB");
-                return plexAccount.GetAuthToken;
-            }
-
-            _log.InformationLine("Plex AuthToken has expired, refreshing Plex AuthToken now");
-
-            // TODO:Account for 2FA
-            // return await _plexApiWrapper.RefreshPlexAuthTokenAsync(plexAccount);
-            await Task.CompletedTask;
-            throw new Exception("Plex AuthToken has expired and needs to be refreshed");
-        }
-
-        return Result.Fail($"PlexAccount with Id: {plexAccount.Id} contained an empty AuthToken!").LogError();
-    }
 }
