+++ src/PlexApi/PlexApiWrapper.cs
@@ -80,71 +80,6 @@ public class PlexApiWrapper
     }
 
     /// <summary>
-    ///     Retrieves all the accessible plex server based on the <see cref="PlexAccount" /> token
-    ///     Including the various unique connections to each server.
-    ///     <remarks>https://plex.tv/api/v2/resources?X-Plex-Token={{AUTH_TOKEN}}</remarks>
-    /// </summary>
-    /// <param name="authToken">The Plex account authentication token.</param>
-    /// <returns> A list of <see cref="PlexDevice" /> with all the connections to the servers.</returns>
-    public async Task<Result<List<PlexDevice>>> GetAccessibleServers(string authToken)
-    {
-        if (string.IsNullOrEmpty(authToken))
-            return ResultExtensions.IsEmpty(nameof(authToken)).LogError();
-
-        var plexTvClient = CreateTvClient(
-            authToken,
-            new PlexApiClientOptions { ConnectionUrl = string.Empty, Timeout = 15 }
-        );
-
-        var result = await Task.WhenAll(
-            ToResponse(plexTvClient.Plex.GetServerResourcesAsync(clientID: GetClientId)),
-            ToResponse(
-                plexTvClient.Plex.GetServerResourcesAsync(
-                    clientID: GetClientId,
-                    includeHttps: IncludeHttps.Enable,
-                    includeRelay: IncludeRelay.Enable,
-                    includeIPv6: IncludeIPv6.Enable
-                )
-            )
-        );
-
-        if (result[0].IsFailed && result[1].IsFailed)
-            return Result.Merge(result[0].ToResult(), result[1].ToResult());
-
-        if (result[0].IsFailed && result[1].IsSuccess)
-            return result[1].ToApiResult(x => x.PlexDevices ?? []);
-
-        if (result[0].IsSuccess && result[1].IsFailed)
-            return result[0].ToApiResult(x => x.PlexDevices ?? []);
-
-        var deviceList1 = result[0].Value?.PlexDevices?.FindAll(x => x.Provides.Contains("server")) ?? [];
-        var deviceList2 = result[1].Value?.PlexDevices?.FindAll(x => x.Provides.Contains("server")) ?? [];
-
-        var uniqueConnections = new HashSet<string>();
-
-        foreach (var device1 in deviceList1)
-        {
-            var device2 = deviceList2.FirstOrDefault(x => x.ClientIdentifier == device1.ClientIdentifier);
-            if (device2 is null || !device2.Connections.Any())
-                continue;
-
-            var serverConnections = device1.Connections.Concat(device2.Connections).ToList();
-
-            device1.Connections.Clear();
-
-            foreach (var connection in serverConnections)
-            {
-                if (uniqueConnections.Add(connection.Uri))
-                {
-                    device1.Connections.Add(connection);
-                }
-            }
-        }
-
-        return Result.Ok(deviceList1);
-    }
-
-    /// <summary>
     ///     Returns a detailed overview of the PlexLibraries in a PlexServer from the PlexAPI.
     ///     <remarks>{{SERVER_URL}}/library/sections?X-Plex-Token={{SERVER_TOKEN}}</remarks>
     /// </summary>
