+++ tests/BaseTests/Common/BaseUnitTest/BaseUnitTest.Common.cs
@@ -71,6 +71,7 @@ public partial class BaseUnitTest
                 // All different dependencies that are needed for the endpoint need to be added here. And then they can be mocked in the test.
                 s.AddTransient(_ => mock.Create<ILog>());
                 s.AddTransient(_ => mock.Create<IPlexRipperDbContext>());
+                s.AddTransient(_ => mock.Create<ICommandDispatch>());
                 s.AddTransient(_ => mock.Create<IPlexApiService>());
                 s.AddSingleton(_ => mock.Create<IMediator>());
                 s.AddSingleton(_ => mock.Create<ISchedulerService>());
